ファイル定義とメトリクス毎の構造体に分解して管理する
